requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
chardet==5.2.0
crispy-bootstrap4==2023.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-crispy-forms==2.1
django-import-export==3.3.6
django-js-asset==2.2.0
django-modeltranslation==0.18.11
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
environ==1.0
environs==10.1.0
et-xmlfile==1.1.0
MarkupPy==1.14
marshmallow==3.20.1
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.0.9
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzlocal==5.2
xlrd==2.0.1
xlwt==1.3.0